+++ controllers/courseCtrl.php
@@ -39,18 +39,29 @@
 			//Permissions validate
 			//ins validate
 			$data = array();
-			$data['name'] = $this->validateName($_GET['name']);
-			$data['email'] = $this->validateEMail($_GET['email']);
-			
-			$status = $this->model->insert($data);
-			
-			if($status){
-				//Load the view for an inserted course
-				include('views/courseInsert.php');
-				//The course has been inserted
-			}
-			else{
-				include('views/error.php');
+			if(empty($_GET['scholar_cycle'])||empty($_GET['name'])||empty($_GET['section'])||empty($_GET['NRC'])||empty($_GET['academy'])||empty($_GET['class_days'])||empty($_GET['hours_per_day'])||empty($_GET['hours_each_day'])||){
+				include'views/error.php';
+			}else{
+				$data['scholar_cycle'] = $this->validateCycle($_GET['scholar_cycle']);
+				$data['name'] = $this->validateName($_GET['name']);
+				$data['section'] = $this->validateSection($_GET['section']);
+				$data['NRC'] = $this->validateNrc($_GET['NRC']);
+				$data['academy'] = $this->validateAcademy($_GET['academy']);
+				$data['class_days'] = $this->validateClass($_GET['class_days']);
+				$data['hours_per_day'] = $this->validateHoursPerDay($_GET['hours_per_day']);
+				$data['hours_each_day'] = $this->validateHoursEachDay($_GET['hours_each_day']);
+
+				$status = $this->validateIns($data);
+				if($status)
+					$status = $this->model->insert($data);
+				if($status){
+					//Load the view for an updated student
+					include('views/courseInsert.php');
+					//The student has been updated
+				}
+				else{
+					include('views/error.php');
+				}
 			}
 		}
 		/**
@@ -60,17 +71,29 @@
 			//Permissions validate
 			//ins validate
 			$data = array();
-			$data['name'] = $this->validateName($_GET['name']);
-			$data['email'] = $this->validateEMail($_GET['email']);
-			$status = $this->model->delete($data);
-			
-			if($status){
-				//Load the view for a deleted course
-				include('views/courseDelete.php');
-				//The course has been deleted
-			}
-			else{
-				include('views/error.php');
+			if(empty($_GET['scholar_cycle'])||empty($_GET['name'])||empty($_GET['section'])||empty($_GET['NRC'])||empty($_GET['academy'])||empty($_GET['class_days'])||empty($_GET['hours_per_day'])||empty($_GET['hours_each_day'])||){
+				include'views/error.php';
+			}else{
+				$data['scholar_cycle'] = $this->validateCycle($_GET['scholar_cycle']);
+				$data['name'] = $this->validateName($_GET['name']);
+				$data['section'] = $this->validateSection($_GET['section']);
+				$data['NRC'] = $this->validateNrc($_GET['NRC']);
+				$data['academy'] = $this->validateAcademy($_GET['academy']);
+				$data['class_days'] = $this->validateClass($_GET['class_days']);
+				$data['hours_per_day'] = $this->validateHoursPerDay($_GET['hours_per_day']);
+				$data['hours_each_day'] = $this->validateHoursEachDay($_GET['hours_each_day']);
+
+				$status = $this->validateIns($data);
+				if($status)
+					$status = $this->model->delete($data);
+				if($status){
+					//Load the view for an updated student
+					include('views/courseDelete.php');
+					//The student has been updated
+				}
+				else{
+					include('views/error.php');
+				}
 			}
 		}
 		/**
@@ -79,16 +102,29 @@
 		function select(){
 			//Permissions validate
 			$data = array();
-			$data['name'] = $this->validateName($_GET['name']);
-			$data['email'] = $this->validateEMail($_GET['email']);
-			$status = $this->model->select($data);
-			
-			if($status){
-				//Load the view for a selected course
-				include('views/courseSelect.php');
-			}
-			else{
-				include('views/error.php');
+			if(empty($_GET['scholar_cycle'])||empty($_GET['name'])||empty($_GET['section'])||empty($_GET['NRC'])||empty($_GET['academy'])||empty($_GET['class_days'])||empty($_GET['hours_per_day'])||empty($_GET['hours_each_day'])||){
+				include'views/error.php';
+			}else{
+				$data['scholar_cycle'] = $this->validateCycle($_GET['scholar_cycle']);
+				$data['name'] = $this->validateName($_GET['name']);
+				$data['section'] = $this->validateSection($_GET['section']);
+				$data['NRC'] = $this->validateNrc($_GET['NRC']);
+				$data['academy'] = $this->validateAcademy($_GET['academy']);
+				$data['class_days'] = $this->validateClass($_GET['class_days']);
+				$data['hours_per_day'] = $this->validateHoursPerDay($_GET['hours_per_day']);
+				$data['hours_each_day'] = $this->validateHoursEachDay($_GET['hours_each_day']);
+
+				$status = $this->validateIns($data);
+				if($status)
+					$status = $this->model->select($data);
+				if($status){
+					//Load the view for an updated student
+					include('views/courseSelect.php');
+					//The student has been updated
+				}
+				else{
+					include('views/error.php');
+				}
 			}
 		}
 		
@@ -98,17 +134,29 @@
 		function update(){
 			//Permissions validate
 			$data = array();
-			$data['name'] = $this->validateName($_GET['name']);
-			$data['email'] = $this->validateEMail($_GET['email']);
-			$status = $this->model->update($data);
-			
-			if($status){
-				//Load the view for a updated course
-				include('views/courseUpdate.php');
+			if(empty($_GET['scholar_cycle'])||empty($_GET['name'])||empty($_GET['section'])||empty($_GET['NRC'])||empty($_GET['academy'])||empty($_GET['class_days'])||empty($_GET['hours_per_day'])||empty($_GET['hours_each_day'])||){
+				include'views/error.php';
+			}else{
+				$data['scholar_cycle'] = $this->validateCycle($_GET['scholar_cycle']);
+				$data['name'] = $this->validateName($_GET['name']);
+				$data['section'] = $this->validateSection($_GET['section']);
+				$data['NRC'] = $this->validateNrc($_GET['NRC']);
+				$data['academy'] = $this->validateAcademy($_GET['academy']);
+				$data['class_days'] = $this->validateClass($_GET['class_days']);
+				$data['hours_per_day'] = $this->validateHoursPerDay($_GET['hours_per_day']);
+				$data['hours_each_day'] = $this->validateHoursEachDay($_GET['hours_each_day']);
+
+				$status = $this->validateIns($data);
+				if($status)
+					$status = $this->model->update($data);
+				if($status){
+					//Load the view for an updated student
+					include('views/courseUpdate.php');
+					//The student has been updated
+				}
+				else{
+					include('views/error.php');
+				}
 			}
-			else{
-				include('views/error.php');
-			}
-		}
 	 }
 ?>
